refactor: Make user_id 64bit
Allows, giving out user_id's like candy without risking exhaustion.

+++ src/tui/modes.rs
@@ -1195,7 +1195,7 @@ mod tests {
             push_room_message(
                 &mut app,
                 index + 1,
-                UserId(index as u32 + 1),
+                UserId(index as u64 + 1),
                 index * 120_000,
                 format!("message {index}"),
             );
@@ -1219,7 +1219,7 @@ mod tests {
             push_room_message(
                 &mut app,
                 index + 1,
-                UserId(index as u32 + 1),
+                UserId(index as u64 + 1),
                 index * 120_000,
                 format!("message {index}"),
             );
